Clean and re-organize modules
